src/main/java/com/example/todolistappspring/services/TaskService.java
package com.example.todolistappspring.services;

import com.example.todolistappspring.models.Task;
import com.example.todolistappspring.models.User;
import com.example.todolistappspring.repositories.TaskRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class TaskService {
    @Autowired
    private TaskRepository taskRepository;

    public List<Task> findAll() {
        return taskRepository.findAll();
    }

    public Task save(User task) {
        return taskRepository.save(task);
    }

    public void deleteById(Long id) {
        taskRepository.deleteById(id);
    }
}

